Refactor/organization cleanup

+++ backend/EduTracker.Application/Models/SessionData.cs
@@ -16,23 +16,26 @@ public sealed record SessionData(
         return now >= ExpiresAt || now >= AbsoluteExpiresAt || IsRevoked;
     }
 
-    public bool ShouldRefresh(double thresholdPercent)
+    public bool ShouldRefresh(float thresholdPercent)
     {
         if (thresholdPercent < 1 || thresholdPercent > 100)
             throw new ArgumentOutOfRangeException(nameof(thresholdPercent), "Threshold percent must be between 1 and 100.");
 
-        double thresholdFraction = thresholdPercent / 100.0;
+        double thresholdFraction = thresholdPercent / 100;
 
-        if (IsRevoked) return false;
+        if (IsRevoked)
+            return false;
 
         DateTime now = DateTime.UtcNow;
 
-        if (now >= AbsoluteExpiresAt) return false;
+        if (now >= AbsoluteExpiresAt)
+            return false;
 
         TimeSpan remaining = ExpiresAt - now;
         TimeSpan total = ExpiresAt - CreatedAt;
 
-        if (total <= TimeSpan.Zero) return true;
+        if (total <= TimeSpan.Zero)
+            return true;
 
         double remainingPercent = remaining.TotalSeconds / total.TotalSeconds;
         return remainingPercent <= thresholdFraction;
